Rename parsers directory to decode
